+++ src/wizard/wizard-footer.component.jsx
@@ -7,17 +7,17 @@ import { Icon } from '../icons/index';
 import './wizard-footer.component.scss';
 
 
-export default class WizardFooter extends React.Component {
+export default class WizardFooter extends React.PureComponent {
 
   static propTypes = {
     save: PropTypes.func.isRequired,
     cancel: PropTypes.func.isRequired,
     localizationTexts: PropTypes.shape({
-      save: PropTypes.string,
-      cancel: PropTypes.cancel,
+      save: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
+      cancel: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
     }).isRequired,
     steps: PropTypes.arrayOf(PropTypes.shape({
-      name: PropTypes.string.isRequired,
+      name: PropTypes.oneOfType([PropTypes.string, PropTypes.element]).isRequired,
       component: PropTypes.node.isRequired,
     })).isRequired,
     selectPage: PropTypes.func.isRequired,
@@ -39,34 +39,41 @@ export default class WizardFooter extends React.Component {
   )
 
   render() {
+    const { currentStep, localizationTexts, showPageIndicator, steps } = this.props;
     return (
       <div id="wizard-footer">
-        <div id="bottom-indicators">
-          {this.props.showPageIndicator && this.getIndicators()}
-        </div>
+        {showPageIndicator &&
+          <div id="bottom-indicators">
+            {this.getIndicators()}
+          </div>
+        }
         <div id="bottom-buttons">
           <Button
+            disabled={currentStep === 0}
             id="previous-step"
-            onClick={(event) => {
-              this.props.selectPage(event, this.props.currentStep - 1);
-            }}
+            onClick={event => this.props.selectPage(event, currentStep - 1)}
           >
             <Icon type="indicator" name="CaretLeft" height={30} width={30} />
           </Button>
           <span className="divider" />
-          <Button id="save-button" onClick={this.props.save}>
-            {this.props.localizationTexts.save}
+          <Button
+            id="save-button"
+            onClick={this.props.save}
+          >
+            {localizationTexts.save}
           </Button>
           <span className="divider-small" />
-          <Button id="cancel-button" onClick={this.props.cancel}>
-            {this.props.localizationTexts.cancel}
+          <Button
+            id="cancel-button"
+            onClick={this.props.cancel}
+          >
+            {localizationTexts.cancel}
           </Button>
           <span className="divider" />
           <Button
+            disabled={currentStep === steps.length - 1}
             id="next-step"
-            onClick={(event) => {
-              this.props.selectPage(event, this.props.currentStep + 1);
-            }}
+            onClick={event => this.props.selectPage(event, currentStep + 1)}
           >
             <Icon type="indicator" name="CaretRight" height={30} width={30} />
           </Button>